Finalize backend, add more error logging

+++ backend/controllers/product.controller.js
@@ -51,12 +51,16 @@ export const updateProducts = async (req, res) => {
 export const deleteProducts = async (req, res) => {
     const { id } = req.params;
 
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404),json({success: false, message: "Invalid Product ID"});
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         console.log("Product removed");
         res.status(200).json({ success: true, message: 'Product removed' });
     } catch(error) {
         console.error("Error in Delete Product: ", error.message);
-        res.status(404).json({success: false, message: "Product not found"});
+        res.status(500).json({success: false, message: "Server error"});
     }
 };
